deploy/ml-model/src/requirements.txt
fastapi==0.109.0
llama-cpp-python>=0.1.72
pandas
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
mlflow
boto3
loguru
